refactor: move quota markdown into control plane
