+++ examples/project/project-agent-adoption-smoke.py
@@ -127,11 +127,11 @@ def main() -> int:
         assert first_quota["todo_write_hint"]["section"] == "User Todo / Owner Review Reading Queue", first_quota
         assert first_quota["todo_write_hint"]["user_gate_command_template"] == (
             f"loopx todo add --goal-id {GOAL_ID} --role user --task-class user_gate "
-            "--blocks-agent <agent-id> --text '<public-safe blocking user/owner decision>'"
+            "--blocks-agent <agent-id> --text '<blocking user decision>'"
         ), first_quota
         assert first_quota["todo_write_hint"]["user_action_command_template"] == (
             f"loopx todo add --goal-id {GOAL_ID} --role user --task-class user_action "
-            "--text '<public-safe non-blocking user todo>'"
+            "--text '<non-blocking user todo>'"
         ), first_quota
         assert first_quota["agent_todo_summary"]["open_count"] == 1, first_quota
         assert first_quota["agent_todo_summary"]["first_open_items"][0]["text"] == AGENT_TODO, first_quota
