+++ loopx/control_plane/quota/states.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+
+QUOTA_STATE_ORDER = (
+    "blocked_health",
+    "operator_gate",
+    "focus_wait",
+    "eligible",
+    "waiting",
+    "throttled",
+    "paused",
+)
